build(app): gate mock + shot behind non-default feature
The seeded MockDriver (mock.rs) and RDB_SHOT window capture (shot.rs) are
design-parity tooling that shipped in every build, ~830 lines the size-
budgeted release never needs. Move the heavy mock body into a mock-feature
submodule (mock/imp.rs) and gate shot + the AnyDriver::Mock arms on it;
mock_mode() stays always-compiled and folds to false without the feature,
so no call site needs cfg. Dev screenshots use the new fe-run-mock target.
